requirements.txt
Django~=5.2.6
djangorestframework~=3.16.1
requests~=2.32.5
pandas~=2.3.2
numpy~=2.3.3
python-dateutil~=2.9.0
psycopg[binary]~=3.2
pytest~=8.3
pytest-django~=4.9
python-dotenv~=1.0